small upd + add package for task
