--- chapter_008/src/main/java/ru/job4j/design/lsp/StoreStrategy.java
@@ -1,8 +1,0 @@
-package ru.job4j.design.lsp;
-
-public interface StoreStrategy {
-
-    boolean accept(Food food, double qualityPercents);
-
-    int size();
-}
