+++ chapter_008/src/main/java/ru/job4j/design/lsp/food/Context.java
@@ -1,4 +1,4 @@
-package ru.job4j.design.lsp;
+package ru.job4j.design.lsp.food;
 
 public class Context {
     private StoreStrategy storeStrategy;
