+++ chapter_008/src/main/java/ru/job4j/design/lsp/food/Shop.java
@@ -1,14 +1,14 @@
-package ru.job4j.design.lsp;
+package ru.job4j.design.lsp.food;
 
 import java.util.ArrayList;
 
-public class Trash implements StoreStrategy {
+public class Shop implements StoreStrategy {
     private ArrayList<Food> store = new ArrayList<>();
 
     @Override
     public boolean accept(Food food, double qualityPercents) {
         var result = false;
-        if (qualityPercents >= 75.0) {
+        if (qualityPercents >= 25.0 && qualityPercents <= 75.0) {
             result = store.add(food);
         }
         return result;
